+++ source/feature-renames.json
@@ -19,7 +19,8 @@
 	"global-discussion-list-filters": "global-conversation-list-filters",
 	"extend-discussion-status-filters": "extend-conversation-status-filters",
 	"discussion-links-on-repo-lists": "conversation-links-on-repo-lists",
-	"sort-issues-by-update-time": "sort-conversations-by-update-time",
+	"sort-issues-by-update-time": "last-update-sort",
+	"sort-conversations-by-update-time": "last-update-sort",
 	"clean-issue-filters": "clean-conversation-filters",
 	"separate-draft-pr-button": "one-click-pr-or-gist",
 	"prevent-pr-commit-link-loss": "prevent-link-loss",
